Get rid of old browser tests in the cucumber directory
