+++ test/browser_tests/cucumber-new/step_definitions/pagination.js
@@ -1,6 +1,7 @@
 const pagination = require( '../../shared_objects/pagination.js' );
-const { When } = require( 'cucumber' );
+const { Then, When } = require( 'cucumber' );
 const chai = require( 'chai' );
+const expect = chai.expect;
 const chaiAsPromised = require( 'chai-as-promised' );
 
 chai.use( chaiAsPromised );
@@ -16,5 +17,14 @@ When( /I enter "(\d)" in the page input field/,
   function( pageNumber ) {
 
     return pagination.pageInput.sendKeys( pageNumber );
+  }
+);
+
+Then( /the page url should contain "(.*)"/,
+  function( urlComponent ) {
+
+    return expect( browser.getCurrentUrl() )
+      .to.eventually
+      .contain( urlComponent );
   }
 );
