--- test/browser_tests/cucumber/step_definitions/filterable-list-control.js
@@ -1,9 +1,0 @@
-const filterableListControl = require(
-  '../../shared_objects/filterable-list-control.js'
-);
-const { When } = require( 'cucumber' );
-
-When( /I (.*) the filterable list control/, function( action ) {
-
-  return filterableListControl[action]();
-} );
